pre-fork initialization capabilities.
1. Added frontend changes to spawn a backend server at registration.
2. Added frontend changes to attach multiple scaled-up worker to the same server.
3. Added backend changes to add the server capabilities to the python worker.
4. Added logs and metrics io remapping.
5. Modfied python worker to behave as a child worker to handle connection.

+++ frontend/server/src/main/java/com/amazonaws/ml/mms/wlm/Model.java
@@ -35,7 +35,14 @@ public class Model {
     private int maxWorkers;
     private int batchSize;
     private int maxBatchDelay;
+    private String preforkInit;
+
+    // Port on which the model server is running
+
+    private AtomicInteger port;
     private ReentrantLock lock;
+
+    private WorkerThread serverThread;
 
     // Total number of subsequent inference request failures
     private AtomicInteger failedInfReqs;
@@ -43,7 +50,7 @@ public class Model {
     // Per worker thread job queue. This separates out the control queue from data queue
     private ConcurrentMap<String, LinkedBlockingDeque<Job>> jobsDb;
 
-    public Model(ModelArchive modelArchive, int queueSize) {
+    public Model(ModelArchive modelArchive, int queueSize, String prefork) {
         this.modelArchive = modelArchive;
         batchSize = 1;
         maxBatchDelay = 100;
@@ -51,7 +58,9 @@ public class Model {
         // Always have a queue for data
         jobsDb.putIfAbsent(DEFAULT_DATA_QUEUE, new LinkedBlockingDeque<>(queueSize));
         failedInfReqs = new AtomicInteger(0);
+        port = new AtomicInteger(-1);
         lock = new ReentrantLock();
+        preforkInit = prefork;
     }
 
     public String getModelName() {
@@ -112,7 +121,7 @@ public class Model {
     }
 
     public void removeJobQueue(String threadId) {
-        if (!threadId.equals(DEFAULT_DATA_QUEUE)) {
+        if (!threadId.equals(DEFAULT_DATA_QUEUE) && jobsDb.containsKey(threadId)) {
             jobsDb.remove(threadId);
         }
     }
@@ -174,11 +183,31 @@ public class Model {
         }
     }
 
+    public int getPort() {
+        return port.get();
+    }
+
+    public void setPort(int port) {
+        this.port.set(port);
+    }
+
     public int incrFailedInfReqs() {
         return failedInfReqs.incrementAndGet();
     }
 
     public void resetFailedInfReqs() {
         failedInfReqs.set(0);
+    }
+
+    public WorkerThread getServerThread() {
+        return serverThread;
+    }
+
+    public void setServerThread(WorkerThread serverThread) {
+        this.serverThread = serverThread;
+    }
+
+    public String getPreforkInit() {
+        return preforkInit;
     }
 }
